refactor(hardware): add platform-specific machine ID implementations
Splits machine ID and system UUID functions into platform-specific files
to properly support testing on non-Linux systems like macOS.

## Changes

- Created system_linux.go with real file reading for Linux systems
- Created system_other.go with mock implementations for non-Linux platforms
- Mock implementation generates deterministic IDs based on hostname
- Removed duplicate functions from nodes.go
- Removed fallback logic as mock always provides values

## Why This Change?

The hardware graph builder tests were failing on macOS because:
1. Linux-specific files (/etc/machine-id, /sys/class/dmi/id/product_uuid) don't exist
2. This caused resource names to be empty, failing assertions

The platform-specific approach ensures:
- Tests pass consistently on all platforms
- Real values are used on Linux systems
- Deterministic mock values are used for development/testing on non-Linux systems
